+++ ReleaseNotes.md
@@ -8,3 +8,4 @@
 - Add MileEnablePerMonitorDialogScaling function.
 - Add MileSetWindowCaptionColorAttribute function.
 - Add MileSetWindowUseImmersiveDarkModeAttribute function.
+- Add MILE_WINDOW_SYSTEM_BACKDROP_TYPE enum.