+++ ReleaseNotes.md
@@ -11,3 +11,4 @@
 - Add MILE_WINDOW_SYSTEM_BACKDROP_TYPE enum.
 - Add MileGetWindowSystemBackdropTypeAttribute function.
 - Add MileSetWindowSystemBackdropTypeAttribute function.
+- Add MILE_PREFERRED_APP_MODE enum.