+++ ReleaseNotes.md
@@ -1,6 +1,6 @@
 ﻿# Mile.Windows.Helpers Release Notes
 
-**Mile.Windows.Helpers 1.0 (1.0.5.0)**
+**Mile.Windows.Helpers 1.0 (1.0.6.0)**
 
 - Provide static library mode.
 - Provide NuGet package.